requirements.txt
playwright>=1.35.0
python-dotenv>=1.0.0
pyyaml>=6.0
google-generativeai>=0.3.0
jinja2>=3.1.2
pytest>=7.3.1
pytest-playwright>=0.4.0
pytest-asyncio>=1.0.0
beautifulsoup4>=4.12.0
aiofiles>=23.0.0
pillow>=10.0.0